refactor: modularize protocol structure
- Move tests to dedicated test.rs files
- Split lib.rs into protocol-specific modules
- Reorganize project structure for better maintainability
- Improve code organization and separation of concerns

This change improves code maintainability and makes the codebase
more modular and easier to test.
