+++ library/src/pairing/mod.rs
@@ -1,0 +1,3 @@
+pub mod pairing;
+#[cfg(test)]
+mod test;
