+++ library/src/protos.rs
@@ -1,0 +1,3 @@
+pub mod derec_proto {
+    include!(concat!(env!("OUT_DIR"), "/org.derecalliance.derec.protobuf.rs"));
+}
